+++ build.py
@@ -92,7 +92,7 @@ def build_macos_for_arch(args, build_config: Config, path_to_build: Path, arch: 
 
     cmake.option('RAV_ABORT_ON_ASSERT', 'ON')
     cmake.option('RAV_BENCHMARKS', 'ON')
-    cmake.option('RAV_DEBUG', 'ON')
+    cmake.option('RAV_ENABLE_DEBUG', 'ON')
 
     if spdlog:
         cmake.option('RAV_ENABLE_SPDLOG', 'ON')
@@ -143,7 +143,7 @@ def build_windows(args, arch, build_config: Config, subfolder: str, spdlog: bool
 
     cmake.option('RAV_ABORT_ON_ASSERT', 'ON')
     cmake.option('RAV_BENCHMARKS', 'ON')
-    cmake.option('RAV_DEBUG', 'ON')
+    cmake.option('RAV_ENABLE_DEBUG', 'ON')
 
     if spdlog:
         cmake.option('RAV_ENABLE_SPDLOG', 'ON')
@@ -175,7 +175,7 @@ def build_linux(args, arch, build_config: Config, subfolder: str, spdlog: bool =
 
     cmake.option('RAV_ABORT_ON_ASSERT', 'ON')
     cmake.option('RAV_BENCHMARKS', 'ON')
-    cmake.option('RAV_DEBUG', 'ON')
+    cmake.option('RAV_ENABLE_DEBUG', 'ON')
 
     if spdlog:
         cmake.option('RAV_ENABLE_SPDLOG', 'ON')
@@ -223,7 +223,7 @@ def build_android(args, arch, build_config: Config, subfolder: str, spdlog: bool
     if spdlog:
         cmake.option('RAV_ENABLE_SPDLOG', 'ON')
 
-    cmake.option('RAV_DEBUG', 'ON')
+    cmake.option('RAV_ENABLE_DEBUG', 'ON')
 
     cmake.configure()
     cmake.build()